requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Gemini API (Google AI Studio)
google-generativeai==0.3.2

# GCP dependencies (optional - only needed if using Vertex AI or BigQuery)
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.39.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
